add authorization

+++ app/controller/items.py
@@ -2,6 +2,7 @@ from flask import Blueprint, jsonify, request
 from flask_jwt import jwt_required
 from app import db
 from ..models.items import ItemModel
+from .users import token_required
 
 items = Blueprint('items', __name__)
 
@@ -21,8 +22,8 @@ def get_items(category_id):
 
 
 @items.route('/categories/<int:category_id>/items', methods=['POST'])
-# @jwt_required()
-def create_item(category_id):
+@token_required
+def create_item(category_id, user_id):
 	"""
 	input:
 		- category_id (int)
@@ -44,7 +45,7 @@ def create_item(category_id):
 	if item:
 		return jsonify({'message': 'existed item'}), 400
 
-	new_item = ItemModel(item_name, item_description, category_id)
+	new_item = ItemModel(item_name, item_description, category_id, user_id)
 	new_item.save_to_db()
 
 	return jsonify({'message': 'created!'}), 201
@@ -62,8 +63,6 @@ def get_item(category_id, item_id):
 	"""
 	data = request.get_json()
 
-	item_id = data['item_id']
-
 	item = db.session.query(ItemModel).filter(ItemModel.id == item_id).first()
 
 	if item:
@@ -73,28 +72,28 @@ def get_item(category_id, item_id):
 
 
 @items.route('/categories/<int:category_id>/items/<int:item_id>', methods=['DELETE'])
-# @jwt_required()
-def delete_item(category_id, item_id):
+@token_required
+def delete_item(category_id, item_id, user_id):
 	"""
 	input: item_id, category_id
 	output: delete from item list
 	"""
-	data = request.get_json()
-
-	item_id = data['item_id']
 
 	item = db.session.query(ItemModel).filter(ItemModel.id == item_id).first()
 
 	if item:
-		item.delete_from_db()
-		return jsonify({"message": "deleted!"}), 200
+		if item.user_id == user_id:
+			item.delete_from_db()
+			return jsonify({"message": "deleted!"}), 200
+		else:
+			return jsonify({"message": "Unauthorized"}), 401
 	else:
 		return jsonify({"message": "item not found"}), 404
 
 
 @items.route('/categories/<int:category_id>/items/<int:item_id>', methods=['PUT'])
-# @jwt_required()
-def edit_item(category_id, item_id):
+@token_required
+def edit_item(category_id, item_id, user_id):
 	"""
 	input: item_id
 	output:
@@ -103,13 +102,15 @@ def edit_item(category_id, item_id):
 	"""
 	data = request.get_json()
 
-	item_id = data['item_id']
 	description = data['description']
 
 	item = db.session.query(ItemModel).filter(ItemModel.id == item_id).first()
 
 	if item:
-		item.description = description
-		return jsonify({"message": "updated!"}), 200
+		if item.user_id == user_id:
+			item.description = description
+			return jsonify({"message": "updated!"}), 200
+		else:
+			return jsonify({"message": "Unauthorized"}), 401
 	else:
 		return jsonify({"message": "item not found"}), 404
